requirements.txt
watchdog
pyyaml
sentence-transformers
numpy